+++ src/main/java/net/thedragonteam/armorplus/package-info.java
@@ -8,3 +8,4 @@
  * @version 1.0
  */
 package net.thedragonteam.armorplus;
+